Optimize extended.hasExtended
The original version generate all the extended view and test if the view is
empty.  That's quite problematic as generating the extended view also call
hasExtended, which also...

Change the code to a new set of method, with safe default.

Note: we have the same problem with hasMap, but that's less costy (but still
not negligeable).